requirements.txt
asgiref==3.8.1
Django==5.0.6
django-appconf==1.0.6
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-imagekit==5.0.0
django-js-asset==2.2.0
django-mptt==0.16.0
pilkit==3.0
pillow==10.3.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
